refactor: split main.js presenter into several - films, filters, search and statistic
