+++ src/main.js
@@ -1,24 +1,16 @@
-import filtersData from './get-filters';
-import Film from './film';
-import FilmPopup from './film-popup';
-import Filter from './filter';
-import Search from './search';
-import Statistic from './statistic';
+import filtersData from './filters/filters-data';
 import API from './api';
 import Provider from './provider';
 import Store from './store';
 
-const HIDDEN_CLASS = `visually-hidden`;
+import {renderFilters} from './filters/filters';
+import {showLoader, hideLoader} from './utils/utils';
+import {renderSearch} from './search/search';
+import {renderFilms, renderTopFilms} from './films/films';
 
-const body = document.querySelector(`body`);
-const main = document.querySelector(`.main`);
-const header = document.querySelector(`.header`);
-const profileContainer = document.querySelector(`.header__profile`);
-const filmsWrapper = document.querySelector(`.films`);
 const filtersContainer = document.querySelector(`.main-navigation`);
 const filmsContainer = document.querySelector(`.films-list .films-list__container`);
 const filmsTopContainers = document.querySelectorAll(`.films-list--extra .films-list__container`);
-const showMoreButton = document.querySelector(`.films-list__show-more`);
 const loadingContainer = document.querySelector(`.films-list__title`);
 
 const AUTHORIZATION = `Basic l76oy54048so925dfdfd0`;
@@ -42,7 +34,7 @@ const store = new Store({key: FILMS_STORE_KEY, storage: localStorage});
  * Create a Provider for synchronize working api and store
  * @type {Provider}
  */
-const provider = new Provider({api, store});
+export const provider = new Provider({api, store});
 
 
 window.addEventListener(`offline`, () => {
@@ -55,295 +47,11 @@ window.addEventListener(`online`, () => {
 });
 
 
-/**
- * Function for render films
- * @param {Node} container
- * @param {Object[]} films
- */
-const renderFilms = (container, films) => {
-  container.innerHTML = ``;
-  const fragment = document.createDocumentFragment();
-
-  films.forEach((film) => {
-    const filmComponent = new Film(film);
-
-    filmComponent.onAddToWatchList = () => {
-      film.userInfo.isGoingToWatch = !film.userInfo.isGoingToWatch;
-
-      provider.updateFilm({id: film.id, data: film.toRAW()})
-        .then((newFilm) => {
-          filmComponent.update(newFilm);
-        });
-    };
-
-    filmComponent.onAddToFavorite = () => {
-      film.userInfo.isFavorite = !film.userInfo.isFavorite;
-
-      provider.updateFilm({id: film.id, data: film.toRAW()})
-        .then((newFilm) => {
-          filmComponent.update(newFilm);
-        });
-    };
-
-    filmComponent.onMarkAsWatched = () => {
-      film.userInfo.isViewed = !film.userInfo.isViewed;
-
-      provider.updateFilm({id: film.id, data: film.toRAW()})
-        .then((newFilm) => {
-          filmComponent.update(newFilm);
-        });
-    };
-
-    filmComponent.onCommentsClick = () => {
-      const filmPopupComponent = new FilmPopup(film);
-
-      filmPopupComponent.onComment = (newComment) => {
-        film.comments.push(newComment);
-
-        provider.updateFilm({id: film.id, data: film.toRAW()})
-          .then((newFilm) => {
-            filmPopupComponent.unblockComments();
-            filmPopupComponent.updateComments(newFilm);
-            filmPopupComponent.element.querySelector(`.film-details__watched-status`).textContent = `Comment added`;
-            filmPopupComponent.toggleRemoveCommentButton();
-          })
-          .catch(() => {
-            filmPopupComponent.shake();
-            filmPopupComponent.errorComments();
-          });
-      };
-
-      filmPopupComponent.onRemoveComment = () => {
-        film.comments.pop();
-
-        provider.updateFilm({id: film.id, data: film.toRAW()})
-          .then((newFilm) => {
-            filmPopupComponent.updateComments(newFilm);
-            filmPopupComponent.element.querySelector(`.film-details__watched-status`).textContent = `Comment deleted`;
-            filmPopupComponent.toggleRemoveCommentButton();
-          });
-      };
-
-      filmPopupComponent.onRatingClick = (newRating) => {
-        film.userInfo.rating = newRating;
-
-        provider.updateFilm({id: film.id, data: film.toRAW()})
-          .then((newFilm) => {
-            filmPopupComponent.unblockRating();
-            filmPopupComponent.updateRating(newFilm);
-          })
-          .catch(() => {
-            filmPopupComponent.shake();
-            filmPopupComponent.errorRating();
-          });
-      };
-
-      filmPopupComponent.onClose = (updatedFilm) => {
-        film = Object.assign(film, updatedFilm);
-
-        provider.updateFilm({id: film.id, data: film.toRAW()})
-          .then((newFilm) => {
-            filmComponent.update(newFilm);
-            body.removeChild(filmPopupComponent.element);
-            filmPopupComponent.unrender();
-          })
-          .catch(() => {
-            filmPopupComponent.shake();
-          });
-      };
-
-      filmPopupComponent.render();
-      body.appendChild(filmPopupComponent.element);
-    };
-
-    fragment.appendChild(filmComponent.render());
-  });
-
-  container.appendChild(fragment);
-};
-
-
-/**
- * Function for render top films
- * @param {NodeListOf} containers
- * @param {Object[]} films
- */
-const renderTopFilms = (containers, films) => {
-  containers.forEach((container) => {
-    renderFilms(container, films);
-  });
-};
-
-
-/**
- * Function for filter films
- * @param {Object[]} films
- * @param {String} filterName
- * @return {Object[]}
- */
-const filterFilms = (films, filterName) => {
-  let filteredFilms = films;
-
-  switch (filterName) {
-    case `all`:
-      filteredFilms = films;
-      break;
-    case `watchlist`:
-      filteredFilms = films.filter((it) => it.userInfo.isGoingToWatch);
-      break;
-    case `history`:
-      filteredFilms = films.filter((it) => it.userInfo.isViewed);
-      break;
-    case `favorites`:
-      filteredFilms = films.filter((it) => it.userInfo.isFavorite);
-      break;
-  }
-
-  return filteredFilms;
-};
-
-
-/**
- * Function for update active filter
- * @param {Object} activeFilter
- * @param {Object[]} filters
- * @return {Object[]}
- */
-const updateActiveFilter = (activeFilter, filters) => {
-  for (const filter of filters) {
-    if (filter.isActive) {
-      filter.isActive = false;
-      break;
-    }
-  }
-  activeFilter.isActive = true;
-  return filters;
-};
-
-
-/**
- * Function for check should 'Show more' button be visible or no
- * @param {Object[]} films
- */
-const checkLoadMoreButton = (films) => {
-  if (films.length === 0) {
-    showMoreButton.classList.add(HIDDEN_CLASS);
-  } else {
-    showMoreButton.classList.remove(HIDDEN_CLASS);
-  }
-};
-
-
-/**
- * Function for render filters
- * @param {Node} container
- * @param {Object} filters
- * @param {Object[]} films
- */
-const renderFilters = (container, filters, films) => {
-  container.innerHTML = ``;
-  const fragment = document.createDocumentFragment();
-
-  filters.forEach((filter) => {
-    const filterComponent = new Filter(filter);
-
-    filterComponent.onFilter = () => {
-      const filterName = filterComponent.filterId;
-
-      filters = updateActiveFilter(filter, filters);
-      filter.isActive = true;
-      filterComponent.update(filter);
-
-      if (filterName === `stats`) {
-        showStatistic(films);
-        filmsWrapper.classList.add(HIDDEN_CLASS);
-      } else {
-        filmsWrapper.classList.remove(HIDDEN_CLASS);
-        hideStatistic();
-
-        const filteredFilms = filterFilms(films, filterName);
-
-        checkLoadMoreButton(filteredFilms);
-        renderFilms(filmsContainer, filteredFilms);
-      }
-
-      renderFilters(container, filters, films);
-    };
-
-    fragment.appendChild(filterComponent.render());
-  });
-
-  container.appendChild(fragment);
-};
-
-
-/**
- * Function for searching film by title
- * @param {Object[]} films
- * @param {String} title
- * @return {Object[]}
- */
-const searchFilm = (films, title) => films.filter((film) => film.filmInfo.title.toLowerCase().indexOf(title.toLowerCase()) !== -1);
-
-
-/**
- * Function for rendering search
- * @param {Object[]} films
- */
-const renderSearch = (films) => {
-  const search = new Search();
-
-  search.onSearch = () => {
-    const filteredFilms = searchFilm(films, search.input.value);
-
-    checkLoadMoreButton(filteredFilms);
-    renderFilms(filmsContainer, filteredFilms);
-  };
-
-  header.insertBefore(search.render(), profileContainer);
-};
-
-
-/**
- * Function for show statistic
- * @param {Object[]} films
- */
-const showStatistic = (films) => {
-  const statisticComponent = new Statistic(films);
-  main.appendChild(statisticComponent.render());
-};
-
-
-/** Function for hide statistic */
-const hideStatistic = () => {
-  if (document.querySelector(`.statistic`)) {
-    main.removeChild(document.querySelector(`.statistic`));
-  }
-};
-
-
-/**
- * Function for show loader
- * @param {String} text
- */
-const showLoader = (text = `Loading movies...`) => {
-  loadingContainer.textContent = text;
-  loadingContainer.classList.remove(HIDDEN_CLASS);
-};
-
-
-/** Function for hide loader */
-const hideLoader = () => {
-  loadingContainer.textContent = `Loading movies...`;
-  loadingContainer.classList.add(HIDDEN_CLASS);
-};
-
-
-showLoader();
+showLoader(loadingContainer);
 
 provider.getFilms()
   .then((films) => {
-    hideLoader();
+    hideLoader(loadingContainer);
     renderSearch(films);
     renderFilms(filmsContainer, films);
     renderTopFilms(filmsTopContainers, films);
