+++ src/utils/utils.js
@@ -14,6 +14,8 @@ export const Time = {
   SECOND: 1000
 };
 
+export const HIDDEN_CLASS = `visually-hidden`;
+
 /**
  * Function for denerate random number, not including max
  * @param {Number} max
@@ -42,4 +44,39 @@ export const createElement = (template) => {
  */
 export const isNumeric = (n) => {
   return !isNaN(parseFloat(n)) && isFinite(n);
+};
+
+
+/**
+ * Function for check should 'Show more' button be visible or no
+ * @param {Node} button
+ * @param {Object[]} films
+ */
+export const checkLoadMoreButton = (button, films) => {
+  if (films.length === 0) {
+    button.classList.add(HIDDEN_CLASS);
+  } else {
+    button.classList.remove(HIDDEN_CLASS);
+  }
+};
+
+
+/**
+ * Function for show loader
+ * @param {Node} container
+ * @param {String} text
+ */
+export const showLoader = (container, text = `Loading movies...`) => {
+  container.textContent = text;
+  container.classList.remove(HIDDEN_CLASS);
+};
+
+
+/**
+ * Function for hide loader
+ * @param {Node} container
+ */
+export const hideLoader = (container) => {
+  container.textContent = `Loading movies...`;
+  container.classList.add(HIDDEN_CLASS);
 };
